Re #30558 Muon Differences Clean up
Created new file for MuonDiff
